Safari-609.1.11.1/LayoutTests/http/tests/adClickAttribution/second-attribution-converted-with-higher-priority-expected.txt

Tests that a second attribution conversion with higher priority replaces an older with lower priority.


Converted Ad Click Attributions:
WebCore::AdClickAttribution 1
Source: 127.0.0.1
Destination: localhost
Campaign ID: 4
Conversion data: 12
Conversion priority: 4
Conversion earliest time to send: Within 24-48 hours
Conversion request sent: false
